Enhances UI and item presentation
Improves the user interface by adding transitions to buttons and updating the styling for success buttons with hover effects and box shadows.

Reduces the size of item images in the store for better layout.

Adds icons for implants, mods, and tools, and includes images to item definitions.

Restructures the store layout to use a more compact grid on larger screens.

Updates tailwind configuration with new colors, animations and hover utility classes.